refactor: navigate 로직 컴포넌트 단으로 분리
